Realize basic functionality

+++ src/com/hashmals/bmigui/Calculator.java
@@ -30,7 +30,7 @@ public class Calculator {
     }
 
     public double calculateMetric() {
-        return weight / Math.pow((height / 100), 2);
+        return ((double)weight / Math.pow(((double)height / 100), 2));
     }
 
     public double calculateImperial() {
